Moved competition cards to external element
